+++ SSMLearn/Spring/Spring-maven/src/main/java/com/zhou/spring/aop/impl/CalculatorImpl.java
@@ -11,22 +11,22 @@ import org.springframework.stereotype.Component;
 public class CalculatorImpl implements Calculator {
 
 	@Override
-	public double add(double i, double j) {
+	public int add(int i, int j) {
 		return i+j;
 	}
 
 	@Override
-	public double sub(double i, double j) {
+	public int sub(int i, int j) {
 		return i-j;
 	}
 
 	@Override
-	public double mul(double i, double j) {
+	public int mul(int i, int j) {
 		return i*j;
 	}
 
 	@Override
-	public double div(double i, double j) {
+	public int div(int i, int j) {
 		return i/j;
 	}
